Writes code to present and dismiss M.O.T buttons when presentModesOfTransport button is pressed. Shifts code from POIInfoViewController file to POIInfoViewController+objc file
