+++ quavi/quavi/Controllers/POI Info/POIInfoViewController+objc.swift
@@ -6,4 +6,77 @@
 //  Copyright © 2020 Sunni Tang. All rights reserved.
 //
 
-import Foundation
+import UIKit
+
+extension POIInfoViewController {
+    //MARK:--@objc func
+     @objc func handlePresentingButton(_ sender:UIButton){
+         toggleButton(button: sender, onImage: UIImage(systemName: "location.fill")!, offImage: UIImage(systemName: "location")!)
+         
+         switch showButtons{
+         case .hide:
+             activateNewTopConstraints()
+             showMapView = true
+             presentButtons()
+             showButtons = .show
+         case .show:
+             activateTopConstraints()
+             presentButtons()
+             showButtons = .hide
+         }
+     }
+     
+     @objc func continueButtonPressed(_ sender: UIButton) {
+         #warning("push to mapVC")
+         //        self.dismiss(animated: true)
+         self.presentingViewController?.presentingViewController?.presentingViewController?.dismiss(animated: true, completion: nil)
+     }
+     
+     @objc func handleFinishButtonPressed(_ sender: UIButton) {
+         let popupFinalVC = POIPopUpFinalViewController()
+         popupFinalVC.modalPresentationStyle = .fullScreen
+         self.present(popupFinalVC, animated: true)
+     }
+     
+     @objc func handlePresentingMLView(_ sender: UIButton){
+         self.showAlert(title: "Coming Soon...", message: "The team is currently working on the feature to allow for an easter egg scavenger hunt ")
+     }
+    
+    //MARK:-- Private func
+    
+    //toggles the image of the buttons when clicked
+      private func toggleButton(button:UIButton, onImage:UIImage, offImage:UIImage) {
+          button.currentImage == offImage ? button.setImage(onImage, for: .normal) : button.setImage(offImage, for: .normal)
+      }
+    
+    private func switchButtonAlpha() {
+           carButton.alpha = carButton.alpha == 1 ? 0: 1
+           walkButton.alpha = walkButton.alpha == 1 ? 0: 1
+           bikeButton.alpha = bikeButton.alpha == 1 ? 0: 1
+       }
+    
+    private func activateNewTopConstraints() {
+           NSLayoutConstraint.deactivate([carButtonTopConstraint!, bikeButtonTopConstraint!, walkButtonTopConstraint!])
+           NSLayoutConstraint.activate([newCarButtonTopConstraint!, newWalkButtonTopConstraint!, newBikeButtonTopConstraint!])
+       }
+    
+    private func changePresentModesOfTransportBackgroundColor() {
+        presentModesOfTransport.backgroundColor = presentModesOfTransport.backgroundColor == .white ? .black : .white
+        presentModesOfTransport.tintColor = presentModesOfTransport.tintColor == .white ? .black : .white
+    }
+    
+    private func presentButtons() {
+        UIView.animate(withDuration: 0.5, delay: 0, options: .curveEaseInOut, animations: {[weak self] in
+            self?.switchButtonAlpha()
+            self?.view.layoutIfNeeded()
+        }) { (change) in
+            self.changePresentModesOfTransportBackgroundColor()
+        }
+    }
+    
+    private func activateTopConstraints() {
+        NSLayoutConstraint.deactivate([newCarButtonTopConstraint!, newWalkButtonTopConstraint!, newBikeButtonTopConstraint!])
+        NSLayoutConstraint.activate([carButtonTopConstraint!, bikeButtonTopConstraint!, walkButtonTopConstraint!])
+    }
+
+}
